Add more content the README
This commit adds more content and also removes the old content inside
the "What I learned" section. It was having a lot of code snippet and it
could be boring to see a lot of code.

Also, this commit adds all the images.

I add some resources as well on the documentation.